refactor: remove conversation ui re-export stubs
