+++ clients/playground-new/src/kas-ui/components/conversation/ConversationArea/AutoScroll.tsx
