+++ clients/playground-new/src/kas-ui/components/conversation/ConversationArea/ConversationContextUsage.tsx
@@ -1,10 +1,10 @@
-import { Tooltip } from "@/components/ui/tooltip";
-import { getModelPricing, type ModelPricing } from "@/models";
-import { useWorkspaceStore } from "@/state/WorkspaceProvider";
+import { Tooltip } from "../../../../components/ui/tooltip";
+import { getModelPricing, type ModelPricing } from "../../../../models";
+import { useWorkspaceStore } from "../../../../state/WorkspaceProvider";
 import { Button, ProgressCircle, Stack, Text } from "@chakra-ui/react";
 import type { UIMessage } from "@pstdio/kas/kas-ui";
 import { useEffect, useState } from "react";
-import { useEstimatedTokens } from "../hooks/useEstimatedTokens";
+import { useEstimatedTokens } from "../../../hooks/useEstimatedTokens";
 
 interface ConversationContextUsageProps {
   messages: UIMessage[];
