+++ clients/playground-new/src/kas-ui/components/conversation/ConversationArea/MessageList.tsx
@@ -1,6 +1,6 @@
-import { MessageContent, MessageRoot } from "@/components/ui/ai-message";
-import { EmptyState } from "@/components/ui/empty-state";
-import { hasCredentials } from "@/state/actions/hasCredentials";
+import { MessageContent, MessageRoot } from "../../primitives/ai-message";
+import { EmptyState } from "../../primitives/empty-state";
+import { hasCredentials } from "../../../../state/actions/hasCredentials";
 import type { UIMessage, ToolInvocation } from "@pstdio/kas/kas-ui";
 import { Box, Button, Link, Text, VStack } from "@chakra-ui/react";
 import { CassetteTapeIcon } from "lucide-react";
