+++ clients/playground-new/src/kas-ui/components/conversation/ConversationArea/index.ts
@@ -1,0 +1,1 @@
+export * from "./ConversationArea";
