+++ clients/playground-new/src/kas-ui/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./conversation";
+export * from "./primitives";
