+++ clients/playground-new/src/kas-ui/components/primitives/ai-conversation.tsx
@@ -1,4 +1,4 @@
-import { getIconComponent } from "@/utils/getIcon";
+import { getIconComponent } from "../../../utils/getIcon";
 import type { HTMLChakraProps, IconButtonProps } from "@chakra-ui/react";
 import { AbsoluteCenter, chakra, IconButton } from "@chakra-ui/react";
 import { useCallback } from "react";
