+++ clients/playground-new/src/kas-ui/components/primitives/diff-editor.tsx
@@ -1,6 +1,6 @@
 import { DiffEditor as MonacoDiffEditor } from "@monaco-editor/react";
 import { useMemo } from "react";
-import { customTheme } from "./code-editor";
+import { customTheme } from "../../../components/ui/code-editor";
 
 interface DiffEditorProps {
   original: string;
