+++ clients/playground-new/src/kas-ui/components/primitives/empty-state.tsx
