+++ clients/playground-new/src/kas-ui/utils/timeline.ts
@@ -1,4 +1,4 @@
-import type { TimelineDoc, TitleSegment } from "@/components/ui/timeline";
+import type { TimelineDoc, TitleSegment } from "../components/primitives/timeline";
 import type { ToolInvocation } from "@pstdio/kas/kas-ui";
 import { buildDiffTitleSegments, buildFileDiffPreviews } from "./diff";
 import { toolTypeToIconName } from "./toolIcon";
