+++ clients/playground-new/tsconfig.app.json
@@ -23,7 +23,8 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true,
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@/kas-ui/*": ["./src/kas-ui/*"]
     }
   },
   "include": ["src"]
